Added support for rounded bar corners and point overlays in range bar…

+++ data/model/src/main/java/com/hdil/saluschart/data/model/model/HeartRate.kt
@@ -11,5 +11,5 @@ data class HeartRate(
 
 data class HeartRateSample(
     val time: Instant,
-    val beatsPerMinute: Long
+    val beatsPerMinute: Int
 )
